+++ lib/src/shared_map_isolate.dart
@@ -150,6 +150,8 @@ class _SharedStoreIsolateMain
   @override
   SharedMap<K, V>? getSharedMap<K, V>(
     String id, {
+    SharedMapEventCallback? onInitialize,
+    SharedMapKeyCallback<K, V>? onAbsent,
     SharedMapEntryCallback<K, V>? onPut,
     SharedMapEntryCallback<K, V>? onRemove,
   }) {
@@ -159,11 +161,24 @@ class _SharedStoreIsolateMain
         throw StateError(
             "[$this] Can't cast `sharedMap` to `SharedMap<$K, $V>`: $sharedMap");
       }
+
+      sharedMap.setCallbacksDynamic<K, V>(
+          onInitialize: onInitialize,
+          onAbsent: onAbsent,
+          onPut: onPut,
+          onRemove: onRemove);
+
       return sharedMap as SharedMap<K, V>;
     }
 
     var o = createSharedMap<K, V>(sharedStore: this, id: id);
-    o.setCallbacksDynamic<K, V>(onPut: onPut, onRemove: onRemove);
+
+    o.setCallbacksDynamic<K, V>(
+        onInitialize: onInitialize,
+        onAbsent: onAbsent,
+        onPut: onPut,
+        onRemove: onRemove);
+
     return o;
   }
 
@@ -249,13 +264,19 @@ class _SharedStoreIsolateAuxiliary
   @override
   FutureOr<SharedMap<K, V>?> getSharedMap<K, V>(
     String id, {
+    SharedMapEventCallback? onInitialize,
+    SharedMapKeyCallback<K, V>? onAbsent,
     SharedMapEntryCallback<K, V>? onPut,
     SharedMapEntryCallback<K, V>? onRemove,
   }) {
     var sharedMap = _sharedMaps[id];
     if (sharedMap != null) {
       var o = sharedMap as SharedMap<K, V>;
-      o.setCallbacksDynamic<K, V>(onPut: onPut, onRemove: onRemove);
+      o.setCallbacksDynamic<K, V>(
+          onInitialize: onInitialize,
+          onAbsent: onAbsent,
+          onPut: onPut,
+          onRemove: onRemove);
       return o;
     }
 
@@ -266,18 +287,22 @@ class _SharedStoreIsolateAuxiliary
     if (sharedMapsPort != null) {
       var ref =
           SharedMapReferenceIsolate(id, sharedStoreReference, sharedMapsPort);
-      return _resolveSharedMapFromReference(ref, onPut, onRemove);
+      return _resolveSharedMapFromReference(
+          ref, onInitialize, onAbsent, onPut, onRemove);
     }
 
     _CallCasted<K, V> callCasted = _buildCallCasted<K, V>();
 
     return sendRequest<SharedMapReference>(
             [_SharedStoreIsolateOperation.getSharedMap, id, callCasted])
-        .then((ref) => _resolveSharedMapFromReference(ref, onPut, onRemove));
+        .then((ref) => _resolveSharedMapFromReference(
+            ref, onInitialize, onAbsent, onPut, onRemove));
   }
 
   FutureOr<SharedMap<K, V>> _resolveSharedMapFromReference<K, V>(
     SharedMapReference? ref,
+    SharedMapEventCallback? onInitialize,
+    SharedMapKeyCallback<K, V>? onAbsent,
     SharedMapEntryCallback<K, V>? onPut,
     SharedMapEntryCallback<K, V>? onRemove,
   ) {
@@ -287,7 +312,13 @@ class _SharedStoreIsolateAuxiliary
     }
 
     var o = createSharedMap<K, V>(sharedReference: ref);
-    o.setCallbacksDynamic<K, V>(onPut: onPut, onRemove: onRemove);
+
+    o.setCallbacksDynamic<K, V>(
+        onInitialize: onInitialize,
+        onAbsent: onAbsent,
+        onPut: onPut,
+        onRemove: onRemove);
+
     return o;
   }
 
@@ -340,6 +371,12 @@ class _SharedStoreIsolateAuxiliary
 mixin _SharedMapIsolate<K, V>
     implements SharedMap<K, V>, SharedObjectIsolate<SharedMapReferenceIsolate> {
   @override
+  SharedMapEventCallback? onInitialize;
+
+  @override
+  SharedMapKeyCallback<K, V>? onAbsent;
+
+  @override
   SharedMapEntryCallback<K, V>? onPut;
 
   @override
@@ -347,8 +384,19 @@ mixin _SharedMapIsolate<K, V>
 
   @override
   void setCallbacks(
-      {SharedMapEntryCallback<K, V>? onPut,
+      {SharedMapEventCallback? onInitialize,
+      SharedMapKeyCallback<K, V>? onAbsent,
+      SharedMapEntryCallback<K, V>? onPut,
       SharedMapEntryCallback<K, V>? onRemove}) {
+    if (onInitialize != null && this.onInitialize == null) {
+      this.onInitialize = onInitialize;
+      onInitialize(this);
+    }
+
+    if (onAbsent != null) {
+      this.onAbsent ??= onAbsent;
+    }
+
     if (onPut != null) {
       this.onPut ??= onPut;
     }
@@ -360,8 +408,19 @@ mixin _SharedMapIsolate<K, V>
 
   @override
   void setCallbacksDynamic<K1, V1>(
-      {SharedMapEntryCallback<K1, V1>? onPut,
+      {SharedMapEventCallback? onInitialize,
+      SharedMapKeyCallback<K1, V1>? onAbsent,
+      SharedMapEntryCallback<K1, V1>? onPut,
       SharedMapEntryCallback<K1, V1>? onRemove}) {
+    if (onInitialize != null && this.onInitialize == null) {
+      this.onInitialize = onInitialize;
+      onInitialize(this);
+    }
+
+    if (onAbsent is SharedMapKeyCallback<K, V>) {
+      this.onAbsent ??= onAbsent as SharedMapKeyCallback<K, V>;
+    }
+
     if (onPut is SharedMapEntryCallback<K, V>) {
       this.onPut ??= onPut as SharedMapEntryCallback<K, V>;
     }
@@ -396,35 +455,36 @@ class _SharedMapIsolateMain<K, V>
     switch (op) {
       case SharedMapOperation.get:
         {
-          var key = args[1];
+          var key = args[1] as K;
           response = get(key);
         }
       case SharedMapOperation.put:
         {
-          var key = args[1];
-          var putValue = args[2];
+          var key = args[1] as K;
+          var putValue = args[2] as V?;
           response = put(key, putValue);
         }
       case SharedMapOperation.putIfAbsent:
         {
-          var key = args[1];
-          var putValue = args[2];
+          var key = args[1] as K;
+          var putValue = args[2] as V?;
           response = putIfAbsent(key, putValue);
         }
       case SharedMapOperation.update:
         {
-          var key = args[1];
+          var key = args[1] as K;
           var updater = args[2];
           response = update(key, updater);
         }
       case SharedMapOperation.remove:
         {
-          var key = args[1];
+          var key = args[1] as K;
           response = remove(key);
         }
       case SharedMapOperation.removeAll:
         {
-          var keys = args[1] as List<K>;
+          var list = args[1] as List;
+          var keys = list is List<K> ? list : list.cast<K>();
           response = removeAll(keys);
         }
       case SharedMapOperation.keys:
@@ -458,7 +518,21 @@ class _SharedMapIsolateMain<K, V>
   }
 
   @override
-  V? get(K key) => _entries[key];
+  V? get(K key) {
+    var value = _entries[key];
+
+    if (value == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        value = onAbsent(key);
+        if (value != null) {
+          _entries[key] = value;
+        }
+      }
+    }
+
+    return value;
+  }
 
   @override
   V? put(K key, V? value) {
@@ -477,6 +551,17 @@ class _SharedMapIsolateMain<K, V>
   @override
   V? putIfAbsent(K key, V? absentValue) {
     var prev = _entries[key];
+
+    if (prev == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        prev = onAbsent(key);
+        if (prev != null) {
+          _entries[key] = prev;
+        }
+      }
+    }
+
     if (prev == null) {
       if (absentValue == null) {
         return null;
@@ -495,6 +580,17 @@ class _SharedMapIsolateMain<K, V>
   @override
   V? update(K key, SharedMapUpdater<K, V> updater) {
     var prev = _entries[key];
+
+    if (prev == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        prev = onAbsent(key);
+        if (prev != null) {
+          _entries[key] = prev;
+        }
+      }
+    }
+
     var value = updater(key, prev);
     return put(key, value);
   }
@@ -502,6 +598,14 @@ class _SharedMapIsolateMain<K, V>
   @override
   V? remove(K key) {
     var v = _entries.remove(key);
+
+    if (v == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        v = onAbsent(key);
+      }
+    }
+
     if (v != null) {
       onRemove.callback(key, v);
     }
@@ -509,8 +613,7 @@ class _SharedMapIsolateMain<K, V>
   }
 
   @override
-  List<V?> removeAll(List<K> keys) =>
-      keys.map((k) => _entries.remove(k)).toList();
+  List<V?> removeAll(List<K> keys) => keys.map(remove).toList();
 
   @override
   List<K> keys() => _entries.keys.toList();